requirements.txt
absl-py==2.1.0
accelerate==0.33.0
adblockparser==0.7
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
airllm==2.10.2
altair==5.5.0
annotated-types==0.7.0
antiorm==1.2.1
anyascii==0.3.2
anyio==4.8.0
apturl==0.5.2
asttokens==3.0.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==24.2.0
audioread==3.0.1
Authlib==1.3.2
autocommand==2.2.2
babel==2.16.0
backports.tarfile==1.2.0
bangla==0.0.2
bark==0.1.5
bcrypt==3.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.35.97
botocore==1.35.97
breadability==0.1.20
Brlapi==0.8.3
cachetools==5.5.2
catalogue==2.0.10
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.3.2
cheroot==10.0.1
CherryPy==18.10.0
cherrypy-cors==1.7.0
click==8.1.8
cloudpathlib==0.19.0
colorama==0.4.4
coloredlogs==15.0.1
command-not-found==0.3
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
cymem==2.0.8
Cython==3.0.11
dataclasses-json==0.6.7
datasets==2.21.0
dateparser==1.1.8
db==0.1.1
db-sqlite3==0.0.1
dbus-python==1.2.18
decorator==5.1.1
defer==1.0.6
diffusers==0.32.2
dill==0.3.8
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnnlib @ https://github.com/podgorskiy/dnnlib/releases/download/0.0.1/dnnlib-0.0.1-py3-none-any.whl
dnntools==0.1.8
dnspython==2.6.1
docopt==0.6.2
docstring_parser==0.16
duplicity==0.8.21
einops==0.8.0
email_validator==2.2.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
encodec==0.1.1
et_xmlfile==2.0.0
eventlet==0.37.0
exceptiongroup==1.2.2
executing==2.1.0
ExifRead==2.3.2
faiss-gpu==1.7.2
fastapi==0.115.8
fastapi-cli==0.0.7
fasteners==0.14.1
ffmpeg==1.4
filelock==3.15.4
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SocketIO==5.3.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
flatbuffers==24.12.23
fonttools==4.55.3
fpdf==1.7.2
frozenlist==1.4.1
fsspec==2024.6.1
funcy==2.0
future==0.18.2
g2pkk==0.1.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-pasta==0.2.0
gpt4all==2.8.2
greenlet==1.1.0
grpcio==1.68.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.12.1
hangul-romanize==0.1.0
httpagentparser==1.9.5
httpcore==1.0.7
httplib2==0.20.2
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
idna==3.3
importlib-metadata==4.6.4
inflect==7.5.0
ipaddr==2.2.0
ipwhois==1.1.0
ipython==8.31.0
itsdangerous==2.2.0
jamo==0.4.1
jaraco.collections==5.1.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jaraco.text==4.0.0
jedi==0.19.2
jeepney==0.7.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kagglehub==0.3.6
keras==3.8.0
keyring==23.5.0
kiwisolver==1.4.8
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.37
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.6
langcodes==3.4.1
langsmith==0.3.10
language-selector==0.1
language_data==1.2.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
legacy==0.1.7
libclang==18.1.1
librosa==0.10.0
llvmlite==0.43.0
lockfile==0.12.2
louis==3.20.0
lxml==4.9.4
macaroonbakery==1.3.1
Mako==1.3.8
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.27.1
netaddr==0.10.1
netifaces==0.11.0
networkx==2.6.3
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
openai==1.63.2
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optimum==1.21.4
optree==0.13.1
orjson==3.10.15
packaging==24.1
pandas==1.5.3
paramiko==2.9.3
parso==0.8.4
peft==0.12.0
pexpect==4.8.0
phonenumbers==8.13.52
photon==1.1.8.3
Pillow==9.0.1
platformdirs==2.5.1
pooch==1.8.2
portend==3.2.0
preshed==3.0.9
prompt_toolkit==3.0.48
protobuf==5.29.2
psutil==6.0.0
ptyprocess==0.7.0
publicsuffixlist==0.9.4
pure_eval==0.2.3
pyaes==1.6.1
pyarrow==17.0.0
pyasn1==0.6.0
pycairo==1.20.1
pycountry==24.6.1
pycparser==2.20
pycups==2.0.1
pydantic==2.9.2
pydantic-settings==2.8.0
pydantic_core==2.23.4
pydeck==0.9.1
pydentic==0.0.1.dev3
pydub==0.25.1
pygexf==0.2.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
pymongo==4.9.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pynndescent==0.5.13
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyPDF2==1.28.6
pypinyin==0.53.0
pyRFC3339==1.1
pysbd==0.3.4
PySocks==1.7.1
python-apt==2.4.0+ubuntu3
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-docx==0.8.11
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.20
python-pptx==0.6.23
python-socketio==5.11.4
python-stdnum==1.20
python-whois==0.7.3
pyttsx3==2.98
pytube==15.0.0
pytz==2024.2
pyxdg==0.27
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.7.24
reportlab==3.6.8
requests==2.32.3
requests-futures==1.0.1
requests-toolbelt==1.0.0
resampy==0.4.3
rich==13.8.0
rich-toolkit==0.13.2
rpds-py==0.23.1
rsa==4.9
s3transfer==0.10.4
safetensors==0.4.4
scikit-learn==1.5.2
scipy==1.11.4
screen-resolution-extra==0.0.0
seaborn==0.13.2
SecretStorage==3.3.1
secure==0.3.0
sentence-transformers==3.1.1
sentencepiece==0.2.0
shellingham==1.5.4
sherlock==0.4.1
sherlock-project==0.15.0
shtab==1.7.1
simple-websocket==1.0.0
six==1.16.0
smart-open==7.0.4
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==1.4.15
srsly==2.4.8
stack-data==0.6.3
starlette==0.45.3
stem==1.8.2
streamlit==1.42.2
Sublist3r==1.0
SudachiDict-core==20241021
SudachiPy==0.6.9
sumy==0.11.0
sympy==1.13.1
systemd-python==234
tdqm==0.0.1
Telethon==1.36.0
tempora==5.7.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
theHarvester==0.0.1
thinc==8.2.5
threadpoolctl==3.5.0
tld==0.13
tokenizers==0.19.1
toml==0.10.2
torch==2.5.1
torchaudio==2.4.1
tornado==6.4.2
tqdm==4.66.5
trainer==0.0.36
traitlets==5.14.3
transformers==4.43.4
triton==3.1.0
trl==0.10.1
TTS==0.22.0
typeguard==4.4.1
typer==0.12.5
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tyro==0.8.10
tzdata==2024.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
umap-learn==0.5.7
unattended-upgrades==0.1
Unidecode==1.3.8
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.34.0
uvloop==0.21.0
virtualenv==20.13.0+ds
wadllib==1.3.6
wasabi==1.1.3
watchdog==6.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
websockets==15.0
Werkzeug==3.1.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==2.3.3
xdg==5
xformers==0.0.29.post1
xkit==0.0.0
XlsxWriter==3.2.0
xxhash==3.5.0
yarl==1.9.6
yt-dlp==2024.12.23
zc.lockfile==3.0.post1
zipp==1.0.0
zstandard==0.23.0
